removed a warning about signed unsigned mismatch
git-svn-id: https://svn.codehaus.org/groovy/trunk/groovy/modules/native_launcher@11308 a5544e8c-8a19-0410-ba12-f9af4593a198

+++ source/jst_paramhandling.c
@@ -79,7 +79,7 @@ extern JstActualParam* jst_processInputParameters( char** args, int numArgs, Jst
           if ( memcmp( paramInfos[ j ].name, arg, len ) == 0 ) {
             found = JNI_TRUE ;
             processedParams[ i ].param = arg ;
-            if ( ( usedSize + len + 1 ) > (int)actualSize ) {
+            if ( (int)( usedSize + len + 1 ) > (int)actualSize ) {
               processedParams = jst_realloc( processedParams, actualSize += ( 100 + len ) ) ;
               if ( !processedParams ) return NULL ;
             }
